refactor: "Apply Presentational and Container Component to ProductsList component."
