+++ src/lessons/usingHooks/types/productsTypes.ts
@@ -1,0 +1,9 @@
+export type ProductT = {
+  id: number;
+  title: string;
+  image: string;
+};
+
+export type ProductPropsT = {
+  product: ProductT;
+};
